refactored everything till ANN.cs

+++ Assets/Scripts/ANN Kart/ANN/Activation_Functions/IActivationFunction.cs
@@ -2,6 +2,6 @@
 {
     public interface IActivationFunction
     {
-        public double CalculateOutput(float weightedSum);
+        public float CalculateOutput(float weightedSum);
     }
 }
